refactor:int型をbigIntegerに統一 ほとんどの場合INT型で十分なんだろうけど、BIGINTが必要になるぐらいビッグなサービスにするぞ！っという意気込みで設計。

+++ src/database/migrations/2020_08_17_021054_商品テーブル作成.php
@@ -15,10 +15,10 @@ class 商品テーブル作成 extends Migration
     public function up()
     {
         Schema::create('商品', function (Blueprint $table) {
-            $table->integer('id');
+            $table->bigInteger('id');
             $table->string('名前');
-            $table->integer('レンタル料金');
-            $table->integer('カテゴリid');
+            $table->bigInteger('レンタル料金');
+            $table->bigInteger('カテゴリid');
             $table->timestamps();
         });
     }
